Created Connection Service
This is the same as the ConnectionHelper in the Six Degrees application. Placed it inside of the services and created connection in startup.cs. This is change is due to it being .Net5. Unsure if this practice is still used or if I need it to consistently be called Helper. I will look into this.